organize plugin source based on current conventions
